+++ index.js
@@ -40,7 +40,7 @@ function main() {
   analyser.respectWhitelist(config.whitelist);
 
   console.log(colors.dim('[3/3]'), '✔️  Done');
-  analyser.logResults(commandOptions.stats);
+  analyser.logResults(commandOptions.stats, commandOptions.occurrences, config.whitelist);
 }
 
 main();