+++ test-apps/ember_lts_2_18/app/components/z-button.js
@@ -1,3 +1,10 @@
 import Component from '@ember/component';
 
 export default Component.extend({});
+
+/*
+  *** TEST CASE - ignore ***
+
+  Testing support for `config.whitelist` (see README.md).
+  This component is meant to be whitelisted and not used anywhere.
+*/