+++ test-apps/ember_lts_2_18_pod/app/modules/components/user/user-signature/component.js
@@ -1,4 +1,11 @@
 import Component from '@ember/component';
 
 export default Component.extend({
+  className: ['user/user-signature'],
 });
+
+/*
+  `className` above is a false friend when it comes to searching for the usage of `user/user-signature` component.
+
+  It should not be counted as an occurrence of a component.
+*/