+++ test-apps/ember_lts_2_18_pod/app/modules/components/y-button/component.js
@@ -1,3 +1,9 @@
 import XButton from 'ember-lts-2-18-pod/modules/components/x-button/component';
 
 export default XButton.extend({});
+
+/*
+  *** TEST CASE - find ***
+
+  Testing support for importing components and extending by other components.
+*/