+++ test-apps/ember_lts_2_18_pod/tests/integration/modules/components/user/user-something/component-test.js
@@ -1,0 +1,24 @@
+import { moduleForComponent, test } from 'ember-qunit';
+import hbs from 'htmlbars-inline-precompile';
+
+moduleForComponent('user/user-something', 'Integration | Component | user/user something', {
+  integration: true
+});
+
+test('it renders', function(assert) {
+  // Set any properties with this.set('myProperty', 'value');
+  // Handle any actions with this.on('myAction', function(val) { ... });
+
+  this.render(hbs`{{user/user-something}}`);
+
+  assert.equal(this.$().text().trim(), '');
+
+  // Template block usage:
+  this.render(hbs`
+    {{#user/user-something}}
+      template block text
+    {{/user/user-something}}
+  `);
+
+  assert.equal(this.$().text().trim(), 'template block text');
+});