+++ test-apps/ember_lts_3_4/app/components/y-button.js
@@ -1,3 +1,9 @@
 import XButton from 'ember-lts-3-4/components/x-button';
 
 export default XButton.extend({});
+
+/*
+  *** TEST CASE - find ***
+
+  Testing support for importing components and extending by other components.
+*/